+++ docs/guide/src/example-plugin/main.go
@@ -1,24 +1,32 @@
 package main
 
 import (
-	"os"
+	"github.com/spf13/cobra"
 
 	"github.com/tendermint/basecoin/cmd/commands"
-	"github.com/urfave/cli"
 )
 
 func main() {
-	//Initialize an instance of basecoin with default basecoin commands
-	app := cli.NewApp()
-	app.Name = "example-plugin"
-	app.Usage = "example-plugin [command] [args...]"
-	app.Version = "0.1.0"
-	app.Commands = []cli.Command{
+
+	//Initialize example-plugin root command
+	var RootCmd = &cobra.Command{
+		Use:   "example-plugin",
+		Short: "example-plugin usage description",
+	}
+
+	//Add the default basecoin commands to the root command
+	RootCmd.AddCommand(
+		commands.InitCmd,
 		commands.StartCmd,
 		commands.TxCmd,
-		commands.KeyCmd,
 		commands.QueryCmd,
+		commands.KeyCmd,
+		commands.VerifyCmd,
+		commands.BlockCmd,
 		commands.AccountCmd,
-	}
-	app.Run(os.Args)
+		commands.UnsafeResetAllCmd,
+	)
+
+	//Run the root command
+	commands.ExecuteWithDebug(RootCmd)
 }