+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.4.0 (TBD)
+
+BREAKING CHANGES:
+
+- CLI now uses Cobra, which forced changes to some of the flag names and orderings
+
+IMPROVEMENTS:
+
+- `basecoin init` doesn't generate error if already initialized
+
 ## 0.3.1 (March 23, 2017)
 
 IMPROVEMENTS: